refactor: auction history persister port pojo interface 추가
