--- src/main/java/until/the/eternity/auctionhistory/domain/dto/external/OpenApiAuctionHistoryResponse.java
@@ -1,15 +1,0 @@
-package until.the.eternity.auctionhistory.domain.dto.external;
-
-import com.fasterxml.jackson.annotation.JsonProperty;
-import java.util.List;
-import until.the.eternity.itemoption.domain.dto.external.OpenApiItemOptionResponse;
-
-public record OpenApiAuctionHistoryResponse(
-        @JsonProperty("item_name") String itemName,
-        @JsonProperty("item_display_name") String itemDisplayName,
-        @JsonProperty("auction_item_category") String itemSubCategory,
-        @JsonProperty("item_count") long itemCount,
-        @JsonProperty("auction_price_per_unit") long auctionPricePerUnit,
-        @JsonProperty("date_auction_buy") String dateAuctionBuy,
-        @JsonProperty("auction_buy_id") String auctionBuyId,
-        @JsonProperty("item_option") List<OpenApiItemOptionResponse> openApiItemOptionResponses) {}
