--- src/main/java/until/the/eternity/auctionhistory/domain/dto/internal/response/AuctionHistoryListResponse.java
@@ -1,7 +1,0 @@
-package until.the.eternity.auctionhistory.domain.dto.internal.response;
-
-import java.util.List;
-
-public record AuctionHistoryListResponse(
-        int count,
-        List<AuctionHistoryDetailResponse<ItemOptionResponse>> auctionHistoryDetailResponses) {}
