+++ src/main/java/until/the/eternity/auctionhistory/domain/service/fetcher/AuctionHistoryFetcher.java
@@ -1,0 +1,9 @@
+package until.the.eternity.auctionhistory.domain.service.fetcher;
+
+import java.util.List;
+import until.the.eternity.auctionhistory.interfaces.external.dto.OpenApiAuctionHistoryResponse;
+import until.the.eternity.common.enums.ItemCategory;
+
+public interface AuctionHistoryFetcher {
+    List<OpenApiAuctionHistoryResponse> fetch(ItemCategory category);
+}
