+++ src/main/java/until/the/eternity/auctionhistory/interfaces/rest/dto/response/AuctionHistoryListResponse.java
@@ -1,0 +1,7 @@
+package until.the.eternity.auctionhistory.interfaces.rest.dto.response;
+
+import java.util.List;
+
+public record AuctionHistoryListResponse(
+        int count,
+        List<AuctionHistoryDetailResponse<ItemOptionResponse>> auctionHistoryDetailResponses) {}
